install: Add support for architecture filtering for kargs
The use case here is that some kernel arguments are architecture
specific, such as the `console=` kargs. This allows a generic base
image to default to including arch-specific values post install.

`EnvProperties` struct to allow for conditional merging of kargs
depending on whether the archiecture matches. Future properties such as
`platform.id` can be added in the future.

Signed-off-by: Luke Yang <luyang@redhat.com>