0.14: Bump c-i-proxy, use external `ocidir` crate
I moved the code that was in `ocidir` to a new repository
and crate: https://github.com/containers/ocidir-rs/

It definitely cleans things up here to use it.

However when doing that I also bumped the `oci-spec`
crate there, which requires a corresponding bump
of the containers-image-proxy crate's semver.

And because we re-export that, we'll need to
bump our semver.

Signed-off-by: Colin Walters <walters@verbum.org>