efi: change `--update-firmware` to match current Anaconda logic
Copy Timothée's comment:
We want to be able to use `bootupd backend install --update-firmware`
 in Anaconda to setup the boot order on EFI systems.

The issue is that when called with `--update-firmware`, bootupd
currently removes the `BootCurrent` boot entry, and then adds a
new BootEntry for the system being installed.

The current Anaconda behavior is to remove all boot entries that
match the product name, then create a new boot entry using the
product name and set it as the first one in the boot order.

To sync with Anaconda behavior, when called with `--update-firmware`,
bootupd will remove all boot entries that match the product name.

See https://github.com/coreos/bootupd/issues/658