build: Install git too
This must be a regression from the timestamp change:
91ed63caf196dd728fb7c7da10bdce40e2008504

Without this we fail to parse the timestamp and get errors;
maybe something else changed.

Of course, we should use a non-Makefile language for
this so we get proper error checking.  I may move some
of the makefile bits into xtask.rs or so.

Signed-off-by: Colin Walters <walters@verbum.org>