install/baseline: Drop aarch64 reserved partition
This was originally done to match what CoreOS is doing, which
is motivated by having Ignition configs that work cross-architecture.

We don't have the same constraints or focus; we encourage
operation at filesystem level in general.

This just changes the baseline `to-disk` installer, anyone who
wants to do things like cross-architecture matching can
of course do so externally.

Signed-off-by: Colin Walters <walters@verbum.org>