Test: implement kargs e2e test
Update check-system.yaml to check the kargs in the installed system.
Update the install Containerfile with install kargs in
`/usr/lib/bootc/install` and day 2 kargs in `/usr/lib/bootc/kargs.d`,
checking to see that the kargs are applied in the installed system.
Update the upgrade Containerfile with different day 2 kargs, checking
to see if the delta is applied properly.

Signed-off-by: Luke Yang <luyang@redhat.com>