deploy: Retrieve bound images when staging new image
This parses any file pointed to by a symlink with a
.container or .image extension found in /usr/lib/bootc/bound-images.d.
An error is thrown if a systemd specifier is found in the parsed
fields. It currently only supports the Image and AuthFile fields.

Some known shortcomings are that each image is pulled synchronously.
It does not do any cleanup during a rollback or if the switch
fails after pulling an image. The install path also needs to
pull bound images.

Co-authored-by: Colin Walters <walters@verbum.org>
Signed-off-by: Chris Kyrouac <ckyrouac@redhat.com>
Signed-off-by: Colin Walters <walters@verbum.org>