install: Use rustix instead of nix in one place
Was looking at cleaning up the blockdev stuff a bit more;
the unsafe nix ioctl code we have is a bit of a wart. Some
searching turns up this use of nix, but rustix exposes
what we want here.

(rustix and nix heavily overlap, I think rustix is a bit
 better designed, though nix has been doing some breaking
 changes to catch up)

Signed-off-by: Colin Walters <walters@verbum.org>