install: Fetch bound images from host's /var/lib/containers
Instead of fetching images from the network at install time,
require them to be present in the host's container storage
(the same place we expect to find the bootc image).

There are numerous advantages to this, such as the basic
fact that this helps one avoid re-pulling images at
each install.

But more importantly it aligns with the needs of "hermetic builds"
which want to pre-fetch content from the network into the target
environment.

For example when generating a disk image with bootc-image-builder
or so, a build tool can consistently preload all the input
images (including the bootc base) into the container store,
and know that the install process to generate the disk
will use those images.

Signed-off-by: Colin Walters <walters@verbum.org>