utils: Add a little CommandRunExt helper trait
When I added the `Task` stuff it wasn't with the idea
that we'd use it for *all* subprocess invocations.

I think in some cases we really just do want a `Command` instance
without the extra wrappering.

Add an implementation (there are many of variants of this out
there in lots of Rust codebases)

This makes sense to use instead of `Task` where we're using
`.quiet()` so that we don't print the description, and especially
where we don't expect the task to fail usually, so it's OK
if the error message is odd.

Signed-off-by: Colin Walters <walters@verbum.org>