cli: Add new `state wipe-ostree` subcommand
Solves #776

Right now the integration tests have some hacky code to clear out
`ostree` stateroots  (see `reset_root` function), it would be better if
they could rely on a bootc command to take care of that for them instead

This commit adds a new `state wipe-ostree` subcommand to the CLI that
removes all `ostree` deployments from the system using native ostree
bindings. This makes it easier to later rm -rf the stateroots and have a
clean slate for the next test.

Signed-off-by: Omer Tuchfeld <omer@tuchfeld.dev>