install: Drop code/test uses of `--security-opt`
We think this is unnecessary now; part of improving
the ergonomics of `bootc install` in general, but
especially with the `to-existing-root` path.

Once this lands, at some point later then we
can also remove it from all of the documentation.
But the most safe thing is to leave it in the
docs for a bit longer.

Closes: https://github.com/containers/bootc/issues/928

Signed-off-by: Colin Walters <walters@verbum.org>