tests: Adapt for newer nushell
It looks like the default `column1` became `column0` which
broke this. But really we don't need a table, we just need a list;
and piping to `from csv` was a weird way to go about splitting
lines into a list.

Change the assertion to just expect a list which simplifies
the code.

Signed-off-by: Colin Walters <walters@verbum.org>