lib: Drop usage of hacky run_in_host_mount_ns
What we're doing here is a bit gross because while we're
changing the mountns, other things (especially networking)
will be inherited from our container, creating an ill-defined
mess. If we do need true host code execution then we should
actually use systemd-run or so.

But we don't need that here - now that we've been automatically
ensuring the `/var/lib/containers` mount, we can use the version
of podman in the container image here.

Signed-off-by: Colin Walters <walters@verbum.org>