WIP: Add progress for install
It turns out to be *really* hard to make this work
with my default dev setup of podman-machine...FIFOs
don't work over virtiofs for unsurprising reasons,
and podman doesn't do fd passing with podman-remote.

I think the best fix is to add pipe fd passing with
podman remote and have it act as a copying proxy.

Signed-off-by: Colin Walters <walters@verbum.org>