mount: factor open_tree_from_pidns()
We might also want to open mount objects in our own mount namespace.
This is particularly useful when recursive is set to `false`.  For
example, we can use this to perform operations directly on the root
filesystem of the container, rather than having to deal with all of the
bind mounts added by the container runtime.

Factor the actual code for calling open_tree() into its own wrapper
function with the same name.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>