cli: run lints on image root filesystem (only)
There are only three operations that make use of the `root` variable
declared at the top of the main CLI `run_from_opt()` function.  Move the
definition of this variable into each of those users.

For "lint", instead of opening "/" and using it for the `root` variable
in the main CLI function, use `open_tree()` to create a non-recursive
mountpoint to operate on instead.

This lets us operate directly on the container image filesystem and not
on the unified view with all the bind mounts set up by the container
runtime.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>