cli: drop support for verifying ostree remotes
The `install` and `switch` verbs supported passing an OSTree remote to
use to verify the embedded OSTree commit. This option in fact did not
function correctly and no verification was actually performace.

While it's possible someone was using this feature, it seems quite
unlikely since the UX is more geared towards native OCI signatures. As a
result, we have decided not to file a CVE for this.

And in fact, since we're planning to keep moving away from ostree,
instead of fixing this bug, just completely rip out support for passing
and OSTree remote.

Note this is distinct from the ostree-ext code, which still does support
signature verification of the embedded OSTree commit. E.g. Fedora
CoreOS is planning to make use of that until it can move to proper OCI
signatures.

While we're here, drop the negation around the container sigpolicy to
make it easier to follow the logic.

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>