bootloader: Always pass full path to bootupd
This will depend on https://github.com/coreos/bootupd/pull/839

And cleans up our bootloader code; move the bootloader
partition definitions back to `install/baseline`. Conceptually
now it's just the `to-disk` flow and bootupd that
know about bootloader partition GUIDs which is how I think
it should be - the bootc core is agnostic to bootloaders.

Signed-off-by: Colin Walters <walters@verbum.org>