storage: Ensure global c/storage is initialized via podman
Two of our tests (and an unknown set of users may) run
`bootc image copy-to-storage` which happens to invoke
skopeo to target /var/lib/containers/storage. If this
happens to be the *first* thing to write to that
c/storage instance it triggers a bug around the podman
image network setup.

Signed-off-by: Colin Walters <walters@verbum.org>