tmpfiles: Support multiple generations
Way back in rpm-ostree's original code here, it always assumed
that it owned The Build.

But in a container-native flow we may have a chain of builds
across an organization and want to generate tmpfiles.d entries
multiple times.

When we do this, we can't just blow away the existing `/usr/lib/tmpfiles.d/bootc-generated.conf`.

Teach the tmpfiles code to have "generations" - we pick a new
file name like `bootc-generated-0.conf`, `bootc-generated-1.conf`
etc.

Also, change the generation code to return a struct with more
info about what it did, which will be used in the linting code.

Signed-off-by: Colin Walters <walters@verbum.org>